requirements.txt
pydantic>=2.0.0
requests>=2.31.0
pandas>=2.1.0
python-dotenv>=1.0.0
openai>=1.0.0
sqlalchemy>=2.0.0
email-validator>=2.0.0
apify-client>=1.0.0
PyYAML>=6.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
jinja2>=3.1.0
httpx>=0.27.0
pytest>=8.0.0
